Area kavramına değinildi.Logger eklendi.Logger a ait ayarlar yapılacaktır. RedireAction ile area ya ulaşamıyoruz.Sınıfça araştırılacaktır

+++ ASMSPresentationLayer/Controllers/NeighbourhoodController.cs
@@ -16,7 +16,7 @@ namespace ASMSPresentationLayer.Controllers
             _neighbourhoodBusinessEngine = neighbourhoodBusinessEngine;
         }
 
-        public JsonResult GetDistrictNeighbourhoods(byte id)
+        public JsonResult GetDistrictNeighbourhoods(int id)
         {
             try
             {
